services/usb-device-xous: abstract AppHID state in a struct
Simpler API, easier to understand and audit.

The UsbDevice and UsbHidClass are stored in AppHID directly, so that callers have less things to worry about: just provide an UsbBusAllocator.

This way we streamline the control flow within main_hw.rs, which is already pretty intricated.

I still have to manage errors from usb-device-xous server to lib.rs, but that will come later.