finalize pre-hash before initializing a new hash
we only have one hash engine in hardware, and we want to use
hardware hashing at boot to speed things up as much as possible
and reduce code size of the bootloader.

This patch causes the pre-hash to finalize before a new hash
is initialized which folds in all the other key parameters.