fix misguided attempt to disambiguate version specifiers
The problem is the original way of specifying crates for remote
download is the same that Rust uses to specify a version of a crate
(e.g. the @ symbol), and so I could not differentiate between the
two cases.

Switch to using a ^ symbol for remote download crates. Hopefully
that doesn't have conflicts??