work around broken AES accelerator in MPW silicon
Haven't dug into it, but the AES instruction is not working
correctly in MPW cramium target silicon. There isn't a great way
to selectively disable hardware acceleration for sub-targets of
patched crates. So, as a work-around, we use xtask to patch
the default arguments of the AES crate, just for cramium-soc
targets, to use software emulated AES. This allows us to move
on with bring-up. To wit, there is another AES accelerator
on-chip that could be much faster than RV32 AES acceleration
(and with better DPA characteristics), and to use that option,
we might have to use a patch like this anyways, because Rust
doesn't have a good way to pass feature flags through patched
crates like this.