handle corner case of OOM during move or lend
In this situation, we enter the OOM with the hardware address
space vs the process space inconsistent (the hardware address
space was transitioned in anticipation of mapping a block into it).

This patch reverts our address space to the original space then
runs the OOM routine, and then puts us back where we were.