add a keyboard hit forwarding service to the console
the logger is the console output. so, unfortunately, this is
where the console input is tied to in hardware.

We want the logger to be 100% dependency free, so we must
rely on a simple "well known server" (e.g. a server with a fixed
address) to bounce keyboard input into the keyboard service itself.

This does mean that anyone who knows the name of that server can
simply inject key presses into the system!